requirements.txt
langchain_huggingface
langchain_chroma
langchain_groq
streamlit
python-dotenv
boto3
sentence-transformers
requests
rank-bm25
numpy